chore: merge types and runtime tests
