+++ test/index.test.ts
@@ -1,8 +1,8 @@
 import './basic'
-import './relationships'
+import './relationships.test'
 import './filters'
-import './resource-embedding'
+import './resource-embedding.test'
 import './transforms'
-import './rpc'
-import './max-affected'
+import './rpc.test'
+import './max-affected.test'
 import './version-and-constants.test'
