+++ test/types/default-inference.test-d.ts
