+++ test/types/index.test-d.ts
@@ -1,10 +1,10 @@
 import { TypeEqual } from 'ts-expect'
 import { expectError, expectType } from 'tsd'
-import { PostgrestClient, PostgrestError } from '../src/index'
-import { Prettify } from '../src/types'
-import { Json } from '../src/select-query-parser/types'
-import { Database } from './types.override'
-import { Database as DatabaseWithOptions } from './types.override-with-options-postgrest13'
+import { PostgrestClient, PostgrestError } from '../../src/index'
+import { Prettify } from '../../src/types'
+import { Json } from '../../src/select-query-parser/types'
+import { Database } from '../types.override'
+import { Database as DatabaseWithOptions } from '../types.override-with-options-postgrest13'
 
 const REST_URL = 'http://localhost:3000'
 const postgrest = new PostgrestClient<Database>(REST_URL)
