+++ test/types/issue-1354-d.ts
