+++ test/types/select-query-parser/types.test-d.ts
